# Add a fast pass for ZString.Join when the argument is a string