Add fast pass when the argument of ZString.Join is a string, similar to string.Join.

## Benchmark result

```csharp    
    public class JoinBenchmark
    {
        string[] Source = new[] { "111", "222", "333" };

        [Benchmark]
        public string StringJoin()
        {
            return string.Join(",", Source);
        }

        [Benchmark]
        public string ZStringJoin() 
        {
            return ZString.Join(",", Source);
        }
    }
```

### Before

```
|      Method |     Mean |    Error |   StdDev |
|------------ |---------:|---------:|---------:|
|  StringJoin | 25.43 ns | 0.458 ns | 0.406 ns |
| ZStringJoin | 56.79 ns | 0.600 ns | 0.531 ns |
```

### After

```
|      Method |     Mean |    Error |   StdDev | Ratio |
|------------ |---------:|---------:|---------:|------:|
|  StringJoin | 26.37 ns | 0.203 ns | 0.190 ns |  1.00 |
| ZStringJoin | 28.57 ns | 0.195 ns | 0.182 ns |  1.08 |
```